bug-fix/implement-defense-feedback
Separate firebase config file from index route file
