+++ Server/Routes/index.js
@@ -1,32 +1,6 @@
-import * as firebase from 'firebase';
-import * as dotenv from 'dotenv';
+import firebase from '../Utilities/config';
 import userRoutes from './users';
 import groupRoutes from './groups';
-
-dotenv.load();
-
-let config = {};
-if (process.env.NODE_ENV === 'test') {
-  config = {
-    apiKey: process.env.APIKEYTEST,
-    authDomain: process.env.AUTHDOMAINTEST,
-    databaseURL: process.env.DATABASEURLTEST,
-    projectId: process.env.PROJECTIDTEST,
-    storageBucket: process.env.STORAGEBUCKETTEST,
-    messagingSenderId: process.env.MESSAGESENDERIDTEST
-  };
-} else {
-  config = {
-    apiKey: process.env.APIKEY,
-    authDomain: process.env.AUTHDOMAIN,
-    databaseURL: process.env.DATABASEURL,
-    projectId: process.env.PROJECTID,
-    storageBucket: process.env.STORAGEBUCKET,
-    messagingSenderId: process.env.MESSAGESENDERID
-  };
-}
-
-firebase.initializeApp(config);
 
 module.exports = (app, io) => {
   // USER API ENDPOINTS
